pyproject.toml
# https://setuptools.pypa.io/en/latest/userguide/quickstart.html
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.interrogate]
ignore-init-method = true
ignore-init-module = false
ignore-magic = false
ignore-semiprivate = false
ignore-private = false
ignore-property-decorators = false
ignore-module = false
fail-under = 100
exclude = ["setup.py", "docs", "build", "examples"]
ignore-regex = ["^get$", "^mock_.*", ".*BaseClass.*"]
verbose = 2
quiet = false
whitelist-regex = []
color = true

[tool.pytest.ini_options]
addopts = "-v --cov --cov-report term-missing"
testpaths = [
    "tests",
]

[tool.isort]
profile = "black"

[tool.ruff.lint]
# See https://docs.astral.sh/ruff/rules/ for a list of rules.
# See https://docs.astral.sh/ruff/configuration/#using-pyprojecttoml for configuration documentation
select = ["E", "F", "I"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages]
find = {}  # Scanning implicit namespaces is active by default

[project]
name = "pyds-cli"
version = "0.5.1"
dependencies = [
    "typer >=0.3.8",
    "pyyaml >=6.0",
    "jinja2 >=3.0.2",
    "loguru >=0.5.3",
    "pyprojroot",
    "python-dotenv >=0.19.1",
    "rich >=10.12.0",
    "ruamel.yaml >=0.17.17",
    "case-converter >=1.0.2",
    "jinja2-strcase >= 0.0.2",
    "cookiecutter",
    "sh",
]
readme = "README.md"

[project.scripts]
pyds = "pyds.cli:app"

[tool.coverage.run]
omit = [
    "*/templates/*",
    "tests/*",
]

[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["osx-arm64", "linux-64"]

[tool.pixi.feature.core.pypi-dependencies]
pyds-cli = { path = ".", editable = true }
case-converter = "*"
jinja2-strcase = "*"

[tool.pixi.feature.core.dependencies]
typer = "*"
pyyaml = "*"
jinja2 = "*"
loguru = "*"
pyprojroot = "*"
python-dotenv = "*"
rich = "*"
ruamel.yaml = "*"
cookiecutter = "*"
sh = "*"

[tool.pixi.feature.testing.dependencies]
pytest = "*"
pytest-cov = "*"

[tool.pixi.feature.docs.dependencies]
mkdocs = "*"
mkdocs-material = "*"
mknotebooks = "*"
mkdocstrings = "*"

[tool.pixi.feature.devtools.pypi-dependencies]
llamabot = ">=0.5.0"
pre-commit = "*"

[tool.pixi.feature.testing.tasks]
test = "pytest"

[tool.pixi.feature.docs.tasks]
build-docs = "mkdocs build"
serve-docs = "mkdocs serve"

[tool.pixi.environments]
default = ["core", "testing", "devtools"]
docs = ["docs"]
testing = ["core", "testing"]
